Apply requested changes and some more clean up.
