+++ src/Avalonia.Diagnostics/Views/EventsView.xaml.cs
@@ -1,4 +1,8 @@
-﻿using System.Linq;
+﻿// Copyright (c) The Avalonia Project. All rights reserved.
+// Licensed under the MIT license. See licence.md file in the project root for full license information.
+
+using System.Linq;
+
 using Avalonia.Controls;
 using Avalonia.Diagnostics.ViewModels;
 using Avalonia.Markup.Xaml;
@@ -7,7 +11,7 @@ namespace Avalonia.Diagnostics.Views
 {
     public class EventsView : UserControl
     {
-        ListBox _events;
+        private ListBox _events;
 
         public EventsView()
         {
